pyproject.toml
[build-system]
requires = ["poetry==1.0.5"]
build-backend = "poetry.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "ticket"
description = "ticket"
version = "0.1.0"
readme = "README.md"
authors = []


[tool.poetry.dependencies]
python = "3.7.7"

# TODO: remove after https://github.com/python-poetry/poetry/issues/2042
importlib-metadata = "<1.2.0,>=1.1.3"

django = "^2.2.11"
django-split-settings = "^1.0"
django-axes = "^5.3"
django-csp = "^3.5"
django-health-check = "^3.11"
django-http-referrer-policy = "^1.0"
django-feature-policy = "^3.1"

psycopg2 = "^2.8"
gunicorn = "^20.0"
python-decouple = "^3.3"
bcrypt = "^3.1"
structlog = "^20.1.0"

# ticket
djangorestframework = "^3.11.0"
django-filter = "^2.2.0"
djangorestframework-filters = "1.0.0.dev0"
django-crispy-forms = "^1.9.0"
pytest-vcr = "^1.0.2"
tqdm = "^4.45.0"
jsonschema = "^3.2.0"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.1"
django-querycount = "^0.7"
django-migration-linter = "^2.1"
django-coverage-plugin = "^1.8"
nplusone = "^1.0"

wemake-python-styleguide = "^0.14"
flake8-pytest-style = "^1.1"
flake8-django = "^0.0.4"
flake8-logging-format = "^0.6.0"
nitpick = "^0.22"

pytest = "^5.4"
pytest-django = "^3.9"
pytest-cov = "^2.8"
# pytest-randomly = "^3.1"
pytest-deadfixtures = "^2.1"
pytest-testmon = "^1.0"
pytest-timeout = "^1.3"
django-test-migrations = "^0.2"
hypothesis = "^5.9"

mypy = "^0.770"
django-stubs = "^1.4"

ipython = "^7.13"

sphinx = "^2.4"
sphinx-autodoc-typehints = "^1.10"
doc8 = "^0.8"

yamllint = "^1.22"
safety = "^1.8"
dotenv-linter = "^0.1"
polint = "^0.4"
dennis = "^0.9"
dump-env = "^1.1"
